Context rework cleanup. Log Warning/Error messages cleaned up.
Changed spawn() to work a little nicer. Will merge subcontexts if the spawner
and spawnee have the same key, then merge the rest with new QUIET flag. Works
for now; may need to switch to an 'ignore these keys' strategy if it doesn't
work for later.

Found a few more places to turn off debug logs.