Skill System, Component, Events.
Added data.codec.adapter module. These are veredi classes that codecs should
use or subclass during/after decoding in order to represent data in game.

Main adapter is 'DataDict', which is a dict which can contain the second thing
'KeyGroup'.

When decoding, a KeyGroup in a dict can be marked for later by a
'KeyGroupMarker' class. This allows a codec to just do its thing and return a
dict that may or may not contain KeyGroupMarker. Then when loaded into a
DataComponent, they are loaded into a DataDict. KeyGroupMarker and their values
are translated into KeyGroups.

All these shenanigans are so that, for example, a SkillComponent can have data
like (yaml snippet):
  !grouped knowledge:
     nature:
       class: false
       ranks: 3
     local:
       class: true
       ranks: 7
   stealth:
     class: true
     ranks: 7

This looks pretty in yaml, and also works as intended in-game. E.g. a
SkillRequest for "knowledge (nature)" is given by the DataDict to teh
'knowledge' KeyGroup, which returns its 'nature' sub-key.

Also have a UserDefined/UserDefinedMarker setup, not really dealt with much in
this commit. Probably needs a tweak or two?

All the Adapter classes use collections.abc in order to explicitly duck-type
themselves.

Added a test for DataDict/KeyGroup.

Changed DataComponent to use DataDict for its mapping structure. Also changed
_verify() to verify the original data. Then _from_data() can cut out stuff it
wants to verify but doesn't need afterwards. E.g.:
  SkillComponent wants to verify the data has a 'skill' key. But it afterwards
  only cares about the value/children of 'skill' in the data.

YAML's component.py has constructor methods and add_constructor() calls for the
!grouped and !user.defined keys. These create our Adapter Markers directly in
the output dict so no need for a post-processing step.

Added a sanity check to registry.get. If it would return a dict, it raises an
exception instead, since it's supposed to retrieve a type or callable.

Skill:

  SkillComponent (DataComponent) for holding on to skill data and returning data
  in a nice API.

  SkillEvents: SkillRequest, SkillResult. For asking SkillSystem for a thing,
  and then getting the thing.

  SkillSystem: the smarts for skill-related things.

Templates:

  Moved my theories for what templates/requirements will eventually be to
  'zest/zata/integration/templates/rules/d20'. Have health and skill currently.

Test Helpers:

  Added zload.py. Has a much more robust take on 'help'. Creates and sets up
  everything SkillSystem's unit test needs in order to load data from test
  file-tree repository into DataComponents.