Null Object Pattern
Added veredi.base.null. In it is the Null object singleton. This can be used to
return a none-ish object that will cascade/coleasce nicely and 'do nothing' when
called with methods the thing it's standing in for should have if it weren't
null right now.

Changed EntityManager and ComponentManager to use Null where appropriate.

Created type hinting NewTypes for Components and Entities:
  - CompOrNull
  - EntityOrNull

Fixed unit tests that tested explicitly for None to test for Null instead.