Created D20RulesSystem base class; converted AbilitySystem.
D20RulesSystem is for, so far, AbilitySystem, SkillSystem, CombatSystem. Moving
all commonality into it from those.

This commit is creation of D20RulesSystem, and conversion of AbilitySystem to be
a subclass of D20RulesSystem.

Changed _query() to also take an EntityId so that it can be used by commands
and non-commands, since they get their EntityIds from different places currently.