WebSocket Mediator test work: disable tests flags
Added way to easily enable/disable tests in __init__.

Added Disabled flags enum. Each member is/should be the name of a test
function (e.g. 'test_ping'). Disabled enum's disabled function can then be used
to check if it's a disabled function by name.

Added method_name() and disabled() functions to testing class.
Added checks to each test to not run if disabled.
Added a test to check the disabled flags and fail if any tests are disabled.

Added checks in the tear-down code to not try to tear down stuff that never ran
because it was disabled.