Changed dotted properties to be class functions.
Encodable needs/wants dotted as a class-level thing, so either it has to be a
function or we have to do shenanigans.

Shenanigans in veredi so fare have been largely annoyances and failures, so...
Changed all dotted to be class methods.

Added note to veredi.base.decorator that /should/ be grep-able for abstract
class method. I never remember which has to go first and they're on different
lines so harder to grep for.

Changed config.registry to do it's auto-dotted thing as class method.

Changed ultra_hyper_debug() to, if first arg is not string, use
pretty.indented() on the object itself, thereby allowing nicely
indented/formatted dicts, etc.

Changed MathSystem to look for veredi.numbers.NumberTypesTuple instead of old
MathValue.