v://0005 LogMixin class for more/easier logger names.
v://0005/2020-12-05T11:06:39-0800/bug/logging-updates

Should be a continuation of Lumberjack's work to increase logging specificity by
having more logger names be the veredi dotted label of the class using it.

Created LogMixin class.
  - Contains a Lumberjack instance (a logging helper class) which will log with
    the dotted name given in `LogMixin._log_config(dotted_str)`.
  - Contains `self._log_*()` functions to match all of Lumberjack's for logging
    at a level or group. Also contains `self._log_will_output()` and other
    helper funcs.

Changed these classes to have/be LogMixin:
  - game.engine
  - game.ecs.base.system
  - Managers:
    - game.ecs.managers (base class)
    - game.ecs.system
    - game.ecs.time
    - game.ecs.event
    - game.ecs.entity
    - game.ecs.component

Changed a bunch of `self._log.*` to `self._log_*` to transition to LogMixin from
a Lumberjack or logger.

Changed a bunch of `log.*` to `self._log_*` to transition to LogMixin from
Lumberjack or logger.

Changed EcsManagers to be an ABC. It already had a func that defaulted to
throwing NotImplementedError, so it should have been already and now:
  - Added `dotted` property to EcsManagers base class to allow set up LogMixin
    in base class.
  - Implemented `dotted` property in all the EcsManager sub-classes.