v://0012 FileBareRepo path init complaint.
Fixed up FileBareRepository init/configuration so that only it checks
for when it should complain, and only its unit-test sets up the 'do not
complain'.

FileBareRepository expects to not have a Configuration most of the time,
so we want a log in there if it gets one unexpectedly, but we don't want
it spamming out for each of the FileBareRepository unit tests.