v://0012 Updated templates.
Updated "__register__.py" to be a more up-to-date and complete example
of registration.

Deleted encodable-specific template - the normal template includes both
our current registrars so it's a good example for multiple now.

Updated "__registrar__.py" to reference the real one, since I don't
expect to have many of those and currently intend to have just the one.